[FIX] restructured files
