+++ .gitignore
@@ -1,0 +1,4 @@
+output.txt
+.idea
+processed
+raw
